changed int to long to calculate download executors

+++ occurrence-download-launcher/src/main/java/org/gbif/occurrence/downloads/launcher/pojo/SparkStaticConfiguration.java
@@ -59,7 +59,7 @@ public class SparkStaticConfiguration {
 
   private Resources executorResources;
   private Resources driverResources;
-  private int recordsPerInstance;
+  private long recordsPerInstance;
   private int minInstances;
   private int maxInstances;
   private int memoryOverheadMb;
